Remove `JvmFnWrapper` implementation, migrate all remaining usages of `Fn` and `FnInitParameters` to `Kotlin lambdas` and `ScopeParameters`, refactor tests, serializers, and related code accordingly.
