+++ lib/src/jvmTest/kotlin/io/wavebeans/lib/stream/ResampleStreamSpec.kt
@@ -20,7 +20,6 @@ class ResampleStreamSpec : DescribeSpec({
     beforeSpec {
         TestWbFileDriver.register()
         WbFileDriver.defaultLocalFileScheme = "test"
-        fnWrapper = JvmFnWrapper()
     }
 
     afterSpec {
